GOTOVO/requirements.txt
python-telegram-bot==20.5
requests==2.31.0
python-dateutil==2.8.2